requirements.txt
streamlit==1.28.0
Pillow==10.1.0

